Adjust number of workers, worker connections and pool size
Dropping the number of workers from 4 to 1 to make it easier to reason
about each instance as to the load that it can handle.

Also reducing the number of connections per worker and making them equal
to the number of database connections results in better utilisation of
CPU and better performance. With the previous configuration we had 256
connections fighting for a maximum of 30 connections.

Finally there is a temporary change of the healthcheck type from `port`
to `process` to allow us to further push the instances without them
being killed because of the load.